requirements.txt
aiosmtplib==3.0.2
annotated-types==0.7.0
anyio==4.6.2.post1
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.5
fastapi-mail==1.4.2
google-api-core==2.23.0
google-auth==2.36.0
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
googleapis-common-protos==1.66.0
grpcio==1.68.0
grpcio-status==1.68.0
h11==0.14.0
idna==3.10
Jinja2==3.1.4
MarkupSafe==3.0.2
proto-plus==1.25.0
protobuf==5.28.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.1
pydantic-settings==2.6.1
pydantic_core==2.27.1
python-dotenv==1.0.1
requests==2.32.3
rsa==4.9
sniffio==1.3.1
starlette==0.41.3
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.1
